Improve plugin log handler
